refactor: move PageSpinner into notices module
